extract the top bar in a new component
