+++ packages/twenty-front/src/modules/command-menu/constants/CommandMenuSearchBarHeight.ts
@@ -1,0 +1,1 @@
+export const COMMAND_MENU_SEARCH_BAR_HEIGHT = 64;
