+++ packages/twenty-front/src/modules/command-menu/constants/CommandMenuSearchBarPadding.ts
@@ -1,0 +1,1 @@
+export const COMMAND_MENU_SEARCH_BAR_PADDING = 3;
